refactor: ScrollToTop 경로 수정
