+++ src/pages/MyPage/MyPage.jsx
@@ -9,7 +9,7 @@ import MySetting from './MySetting';
 import { IoIosArrowDropright } from "react-icons/io";
 import { IoCamera } from "react-icons/io5";
 
-import ScrollToTop from '../../helps/ScrollToTop';
+import ScrollToTop from '../../helpers/ScrollToTop';
 
 const MyPage = () => {
     const [image,setImage] = useState("https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png");
